Move classes from convert.py to new angles.py module, extend functionality
