+++ geodepy/convert.py
@@ -9,596 +9,11 @@ from math import (sin, cos, atan2, radians, degrees,
                   sqrt, cosh, sinh, tan, atan, log)
 import datetime
 from geodepy.constants import utm, grs80
+from geodepy.angles import (DECAngle, DMSAngle, DDMAngle, HPAngle,
+                            hp2dec, hp2dms, hp2ddm, dec2hp, dec2dms, dec2ddm)
 
 # Universal Transverse Mercator Projection Parameters
 proj = utm
-
-
-class DECAngle(float):
-    """
-    Class for working with angles in Decimal Degrees
-    Note: GeodePy also supports working with angles in Decimal Degrees as floats
-    """
-
-    def __init__(self, dec_angle=0.0):
-        """
-        :param dec_angle: float Decimal Degrees angle
-        """
-        super().__init__()
-        self.dec_angle = float(dec_angle)
-
-    def __repr__(self):
-        if self.dec_angle >= 0:
-            return '{DECAngle: +' + str(self.dec_angle) + '}'
-        else:  # negative
-            return '{DECAngle: ' + str(self.dec_angle) + '}'
-
-    # TODO: Everything Else
-
-
-class DMSAngle(object):
-    """
-    Class for working with angles in Degrees, Minutes and Seconds format
-    """
-    def __init__(self, degree, minute=0, second=0.0):
-        """
-        :param degree: Angle: whole degrees component (floats truncated)
-                       Alt: formatted string '±DDD MM SS.SSS'
-        :param minute: Angle: whole minutes component (floats truncated)
-        :param second: Angle: seconds component (floats preserved)
-        """
-        # Convert formatted string 'DDD MM SS.SSS' to DMSAngle
-        if type(degree) == str:
-            str_pts = degree.split(' ')
-            degree = int(str_pts[0])
-            minute = int(str_pts[1])
-            second = float(str_pts[2])
-        # Set sign of object based on sign of any variable
-        if degree == 0:
-            if str(degree)[0] == '-':
-                self.sign = -1
-            elif minute < 0:
-                self.sign = -1
-            elif second < 0:
-                self.sign = -1
-            else:
-                self.sign = 1
-        elif degree > 0:
-            self.sign = 1
-        else:  # degree < 0
-            self.sign = -1
-        self.degree = abs(int(degree))
-        self.minute = abs(int(minute))
-        self.second = abs(second)
-
-    def __repr__(self):
-        if self.sign == -1:
-            signsymbol = '-'
-        elif self.sign == 1:
-            signsymbol = '+'
-        else:
-            signsymbol = 'error'
-        return '{DMSAngle: ' + signsymbol + str(self.degree) + 'd ' +\
-               str(self.minute) + 'm ' + str(self.second) + 's}'
-
-    def __add__(self, other):
-        try:
-            return dec2dms(self.dec() + other.dec())
-        except AttributeError:
-            raise TypeError('Can only add Angle objects with .dec() method '
-                            'together')
-
-    def __radd__(self, other):
-        try:
-            return dec2dms(other.dec() + self.dec())
-        except AttributeError:
-            raise TypeError('Can only add Angle objects with .dec() method '
-                            'together')
-
-    def __sub__(self, other):
-        try:
-            return dec2dms(self.dec() - other.dec())
-        except AttributeError:
-            raise TypeError('Can only subtract Angle objects with .dec() method'
-                            ' together')
-
-    def __rsub__(self, other):
-        try:
-            return dec2dms(other.dec() - self.dec())
-        except AttributeError:
-            raise TypeError('Can only subtract Angle objects with .dec() method'
-                            ' together')
-
-    def __mul__(self, other):
-        try:
-            return dec2dms(self.dec() * other)
-        except TypeError:
-            raise TypeError('Multiply only defined between DMSAngle Object '
-                            'and Int or Float')
-
-    def __rmul__(self, other):
-        try:
-            return dec2dms(other * self.dec())
-        except TypeError:
-            raise TypeError('Multiply only defined between DMSAngle Object '
-                            'and Int or Float')
-
-    def __truediv__(self, other):
-        try:
-            return dec2dms(self.dec() / other)
-        except TypeError:
-            raise TypeError('Division only defined between DMSAngle Object '
-                            'and Int or Float')
-
-    def __abs__(self):
-        return DMSAngle(self.degree, self.minute, self.second)
-
-    def __neg__(self):
-        if self.sign == 1:
-            return DMSAngle(-self.degree, -self.minute, -self.second)
-        else:  # sign == -1
-            return DMSAngle(self.degree, self.minute, self.second)
-
-    def __eq__(self, other):
-        return self.dec() == other.dec()
-
-    def __ne__(self, other):
-        return self.dec() != other.dec()
-
-    def __lt__(self, other):
-        return self.dec() < other.dec()
-
-    def __gt__(self, other):
-        return self.dec() > other.dec()
-
-    def __str__(self):
-        return (str(self.sign * self.degree) + ' ' + str(self.minute) + ' '
-                + str(self.second))
-
-    def rad(self):
-        """
-        Convert to Radians
-        :return: Radians
-        :rtype: float
-        """
-        return radians(self.dec())
-
-    def dec(self):
-        """
-        Convert to Decimal Degrees (float)
-        :return: Decimal Degrees
-        :rtype: float
-        """
-        return self.sign * (self.degree + (self.minute / 60) +
-                            (self.second / 3600))
-
-    def deca(self):
-        """
-        Convert to Decimal Degrees (class)
-        :return: Decimal Degrees
-        :rtype: DECAngle
-        """
-        # return DECAngle(self.dec())
-        pass
-
-    def hp(self):
-        """
-        Convert to HP Notation (float)
-        :return: HP Notation (DDD.MMSSSS)
-        :rtype: float
-        """
-        return self.sign * (self.degree + (self.minute / 100) +
-                            (self.second / 10000))
-
-    def hpa(self):
-        """
-        Convert to HP Notation (class)
-        :return: HP Notation (DDD.MMSSSS)
-        :rtype: HPAngle
-        """
-        return HPAngle(self.hp())
-
-    def ddm(self):
-        """
-        Convert to Degrees, Decimal Minutes Object
-        :return: Degrees, Decimal Minutes Object
-        :rtype: DDMAngle
-        """
-        if self.sign == 1:
-            return DDMAngle(self.degree, self.minute + (self.second/60))
-        else:  # sign == -1
-            return -DDMAngle(self.degree, self.minute + (self.second/60))
-
-
-class DDMAngle(object):
-    """
-    Class for working with angles in Degrees, Decimal Minutes format
-    """
-    def __init__(self, degree, minute=0.0):
-        """
-        :param degree: Angle: whole degrees component (floats truncated)
-        :param minute: Angle:minutes component (floats preserved)
-        """
-        # Convert formatted string 'DDD MM.MMMM' to DDMAngle
-        if type(degree) == str:
-            str_pts = degree.split(' ')
-            degree = int(str_pts[0])
-            minute = int(str_pts[1])
-        # Set sign of object based on sign of any variable
-        if degree == 0:
-            if str(degree)[0] == '-':
-                self.sign = -1
-            elif minute < 0:
-                self.sign = -1
-            else:
-                self.sign = 1
-        elif degree > 0:
-            self.sign = 1
-        else:  # degree < 0
-            self.sign = -1
-        self.degree = abs(int(degree))
-        self.minute = abs(minute)
-
-    def __repr__(self):
-        if self.sign == -1:
-            signsymbol = '-'
-        elif self.sign == 1:
-            signsymbol = '+'
-        else:
-            signsymbol = 'error'
-        return '{DDMAngle: ' + signsymbol + str(self.degree) + 'd ' + \
-               str(self.minute) + 'm}'
-
-    def __add__(self, other):
-        try:
-            return dec2ddm(self.dec() + other.dec())
-        except AttributeError:
-            raise TypeError('Can only add Angle objects with .dec() method '
-                            'together')
-
-    def __radd__(self, other):
-        try:
-            return dec2ddm(other.dec() + self.dec())
-        except AttributeError:
-            raise TypeError('Can only add Angle objects with .dec() method '
-                            'together')
-
-    def __sub__(self, other):
-        try:
-            return dec2ddm(self.dec() - other.dec())
-        except AttributeError:
-            raise TypeError('Can only subtract Angle objects with .dec() method'
-                            ' together')
-
-    def __rsub__(self, other):
-        try:
-            return dec2ddm(other.dec() - self.dec())
-        except AttributeError:
-            raise TypeError('Can only subtract Angle objects with .dec() method'
-                            ' together')
-
-    def __mul__(self, other):
-        try:
-            return dec2ddm(self.dec() * other)
-        except TypeError:
-            raise TypeError('Multiply only defined between DMSAngle Object '
-                            'and Int or Float')
-
-    def __rmul__(self, other):
-        try:
-            return dec2ddm(other * self.dec())
-        except TypeError:
-            raise TypeError('Multiply only defined between DMSAngle Object '
-                            'and Int or Float')
-
-    def __truediv__(self, other):
-        try:
-            return dec2ddm(self.dec() / other)
-        except TypeError:
-            raise TypeError('Division only defined between DMSAngle Object '
-                            'and Int or Float')
-
-    def __abs__(self):
-        return DDMAngle(self.degree, self.minute)
-
-    def __neg__(self):
-        if self.sign == 1:
-            return DDMAngle(-self.degree, -self.minute)
-        else:  # sign == -1
-            return DDMAngle(self.degree, self.minute)
-
-    def __eq__(self, other):
-        return self.dec() == other.dec()
-
-    def __ne__(self, other):
-        return self.dec() != other.dec()
-
-    def __lt__(self, other):
-        return self.dec() < other.dec()
-
-    def __gt__(self, other):
-        return self.dec() > other.dec()
-
-    def __str__(self):
-        return str(self.sign * self.degree) + ' ' + str(self.minute)
-
-    def rad(self):
-        """
-        Convert to Radians
-        :return: Radians
-        :rtype: float
-        """
-        return radians(self.dec())
-
-    def dec(self):
-        """
-        Convert to Decimal Degrees (float)
-        :return: Decimal Degrees
-        :rtype: float
-        """
-        return self.sign * (self.degree + (self.minute / 60))
-
-    def deca(self):
-        """
-        Convert to Decimal Degrees (class)
-        :return: Decimal Degrees
-        :rtype: DECAngle
-        """
-        # return DECAngle(self.dec())
-        pass
-
-    def hp(self):
-        """
-        Convert to HP Notation (float)
-        :return: HP Notation (DDD.MMSSSS)
-        :rtype: float
-        """
-        minute_int, second = divmod(self.minute, 1)
-        return self.sign * (self.degree + (minute_int / 100) +
-                            (second * 0.006))
-
-    def hpa(self):
-        """
-        Convert to HP Notation (class)
-        :return: HP Notation (DDD.MMSSSS)
-        :rtype: HPAngle
-        """
-        return HPAngle(self.hp())
-
-    def dms(self):
-        """
-        Convert to Degrees, Minutes, Seconds Object
-        :return: Degrees, Minutes, Seconds Object
-        :rtype: DMSAngle
-        """
-        minute_int, second = divmod(self.minute, 1)
-        return self.sign * DMSAngle(self.degree, int(minute_int), second * 60)
-
-
-class HPAngle(object):
-    """
-    Class for working with angles in Hewlett-Packard (HP) format
-    """
-    def __init__(self, hp_angle=0.0):
-        """
-        :param hp_angle: float HP angle
-        """
-        self.hp_angle = float(hp_angle)
-
-    def __repr__(self):
-        if self.hp_angle >= 0:
-            return '{HPAngle: +' + str(self.hp_angle) + '}'
-        else:  # negative
-            return '{HPAngle: ' + str(self.hp_angle) + '}'
-
-    def __add__(self, other):
-        try:
-            return HPAngle(dec2hp(self.dec() + other.dec()))
-        except AttributeError:
-            raise TypeError('Can only add Angle objects with .dec() method '
-                            'together')
-
-    def __radd__(self, other):
-        try:
-            return HPAngle(dec2hp(other.dec() + self.dec()))
-        except AttributeError:
-            raise TypeError('Can only add Angle objects with .dec() method '
-                            'together')
-
-    def __sub__(self, other):
-        try:
-            return HPAngle(dec2hp(self.dec() - other.dec()))
-        except AttributeError:
-            raise TypeError('Can only subtract Angle objects with .dec() method'
-                            ' together')
-
-    def __rsub__(self, other):
-        try:
-            return HPAngle(dec2hp(other.dec() - self.dec()))
-        except AttributeError:
-            raise TypeError('Can only subtract Angle objects with .dec() method'
-                            ' together')
-
-    def __mul__(self, other):
-        try:
-            return HPAngle(dec2hp(self.dec() * other))
-        except TypeError:
-            raise TypeError('Multiply only defined between Angle objects and '
-                            'Int or Float')
-
-    def __rmul__(self, other):
-        try:
-            return HPAngle(dec2hp(other * self.dec()))
-        except TypeError:
-            raise TypeError('Multiply only defined between Angle objects and '
-                            'Int or Float')
-
-    def __truediv__(self, other):
-        try:
-            return HPAngle(dec2hp(self.dec() / other))
-        except TypeError:
-            raise TypeError('Division only defined between HPAngle objects '
-                            'and Int or Float')
-
-    def __abs__(self):
-        return HPAngle(abs(self.hp_angle))
-
-    def __neg__(self):
-        return HPAngle(self.hp_angle.__neg__())
-
-    def __eq__(self, other):
-        return self.dec() == other.dec()
-
-    def __ne__(self, other):
-        return self.dec() != other.dec()
-
-    def __lt__(self, other):
-        return self.dec() < other.dec()
-
-    def __gt__(self, other):
-        return self.dec() > other.dec()
-
-    def __int__(self):
-        return int(self.hp_angle)
-
-    def __float__(self):
-        return float(self.hp_angle)
-
-    def __str__(self):
-        return str(self.hp_angle)
-
-    def rad(self):
-        """
-        Convert to Radians
-        :return: Radians
-        :rtype: float
-        """
-        return radians(hp2dec(self.hp_angle))
-
-    def dec(self):
-        """
-        Convert to Decimal Degrees (float)
-        :return: Decimal Degrees
-        :rtype: float
-        """
-        return hp2dec(self.hp_angle)
-
-    def deca(self):
-        """
-        Convert to Decimal Degrees (class)
-        :return: Decimal Degrees
-        :rtype: DECAngle
-        """
-        # return DECAngle(self.dec())
-        pass
-
-    def hp(self):
-        """
-        Convert to HP Notation (float)
-        :return: HP Notation (DDD.MMSSSS)
-        :rtype: float
-        """
-        return float(self.hp_angle)
-
-    def dms(self):
-        """
-        Convert to Degrees, Minutes, Seconds Object
-        :return: Degrees, Minutes, Seconds Object
-        :rtype: DMSAngle
-        """
-        return hp2dms(self.hp_angle)
-
-    def ddm(self):
-        """
-        Convert to Degrees, Decimal Minutes Object
-        :return: Degrees, Decimal Minutes Object
-        :rtype: DDMAngle
-        """
-        return hp2ddm(self.hp_angle)
-
-
-def dec2hp(dec):
-    """
-    Converts Decimal Degrees to HP Notation (float)
-    :param dec: Decimal Degrees
-    :type dec: float
-    :return: HP Notation (DDD.MMSSSS)
-    :rtype: float
-    """
-    minute, second = divmod(abs(dec) * 3600, 60)
-    degree, minute = divmod(minute, 60)
-    hp = degree + (minute / 100) + (second / 10000)
-    return hp if dec >= 0 else -hp
-
-
-def hp2dec(hp):
-    """
-    Converts HP Notation to Decimal Degrees
-    :param hp: HP Notation (DDD.MMSSSS)
-    :type hp: float
-    :return: Decimal Degrees
-    :rtype: float
-    """
-    degmin, second = divmod(abs(hp) * 1000, 10)
-    degree, minute = divmod(degmin, 100)
-    dec = degree + (minute / 60) + (second / 360)
-    return dec if hp >= 0 else -dec
-
-
-def dec2dms(dec):
-    """
-    Converts Decimal Degrees to Degrees, Minutes, Seconds Object
-    :param dec: Decimal Degrees
-    :type dec: float
-    :return: Degrees, Minutes, Seconds Object
-    :rtype: DMSAngle
-    """
-    minute, second = divmod(abs(dec) * 3600, 60)
-    degree, minute = divmod(minute, 60)
-    return (DMSAngle(degree, minute, second) if dec >= 0
-            else DMSAngle(-degree, minute, second))
-
-
-def dec2ddm(dec):
-    """
-    Converts Decimal Degrees to Degrees, Decimal Minutes Object
-    :param dec: Decimal Degrees
-    :type dec: float
-    :return: Degrees, Decimal Minutes Object
-    :rtype: DDMAngle
-    """
-    minute, second = divmod(abs(dec) * 3600, 60)
-    degree, minute = divmod(minute, 60)
-    minute = minute + (second / 60)
-    return DDMAngle(degree, minute) if dec >= 0 else DDMAngle(-degree, minute)
-
-
-def hp2dms(hp):
-    """
-    Converts HP Notation to Degrees, Minutes, Seconds Object
-    :param hp: HP Notation (DDD.MMSSSS)
-    :type hp: float
-    :return: Degrees, Minutes, Seconds Object
-    :rtype: DMSAngle
-    """
-    degmin, second = divmod(abs(hp) * 1000, 10)
-    degree, minute = divmod(degmin, 100)
-    return (DMSAngle(degree, minute, second * 10) if hp >= 0
-            else DMSAngle(-degree, minute, second * 10))
-
-
-def hp2ddm(hp):
-    """
-    Converts HP Notation to Degrees, Decimal Minutes Object
-    :param hp: HP Notation (DDD.MMSSSS)
-    :type hp: float
-    :return: Degrees, Decimal Minutes Object
-    :rtype: DDMAngle
-    """
-    degmin, second = divmod(abs(hp) * 1000, 10)
-    degree, minute = divmod(degmin, 100)
-    minute = minute + (second / 6)
-    return DDMAngle(degree, minute) if hp >= 0 else DDMAngle(-degree, minute)
 
 
 def polar2rect(r, theta):
